New '--make-summary-multiline' option
This is not recommended by the PEP 257, but there are some project that
use it in this way. So, a new '--make-summary-multiline' option is added
to allow those project to be able to use the rest of the docformatter
features.

When this option is set as True and use docformatter over a function
like this one:

def function():
    """This is a one line summary that will be converted to multiline."""

the result will be:

def function():
    """
    This is a one line summary that will be converted to multiline.
    """