Adds support for argument lists like:
Args:
  x: the first argument
  y: the second argument

Currently the system works for these already, kind of.  If the docstring
looks like:

def f(x, y):
    """The identity function.

    Returns what you give it.

    Args:
      x: the argument
      y: ignored
    """
    return x

then docformatter does the right thing.  But if the comment looks like:

def f(x, y):
    """The identity function.

    Args:
      x: the argument
      y: ignored
    """
    return x

then docformatter suggests

def f(x):
    """The identity function.

    Args: x: the argument y: ignored
    """
    return x

Which it seems like is a bug. This updates the code to work properly in
this case.